# api/openapi.yaml
openapi: 3.1.0
info:
  title: FootGrid Core API
  version: 0.1.0
  description: Core vertical-slice contract. Cognito JWT authentication is required outside local development.
servers:
  - url: https://api.footgrid.example/v1
security:
  - bearerAuth: []
paths:
...
  /matches/{matchId}:
    parameters: [{ $ref: '#/components/parameters/MatchID' }]
    get:
      operationId: getMatch
      responses:
        '200': { description: Match, content: { application/json: { schema: { $ref: '#/components/schemas/Match' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /matches/{matchId}/events:
    parameters: [{ $ref: '#/components/parameters/MatchID' }]
    get:
      operationId: listEvents
      parameters:
        - name: after_sequence
          in: query
          schema: { type: integer, minimum: 0 }
      responses:
        '200': { description: Ordered ledger, content: { application/json: { schema: { $ref: '#/components/schemas/EventList' } } } }
    post:
      operationId: appendEvent
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AppendEventRequest' } } }
      responses:
        '201': { description: Event committed, content: { application/json: { schema: { $ref: '#/components/schemas/AppendEventResponse' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Validation' }
...
  /matches/{matchId}/snapshot:
    parameters: [{ $ref: '#/components/parameters/MatchID' }]
    get:
      operationId: getSnapshot
      responses:
        '200': { description: Current authoritative state, content: { application/json: { schema: { $ref: '#/components/schemas/Snapshot' } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    MatchID:
      name: matchId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
  responses:
...
    NotFound:
      description: Resource not found
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Conflict:
      description: Match sequence or status conflict
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ConflictProblem' } } }
    Validation:
      description: Domain validation error
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  schemas:
...
    Problem:
      type: object
      required: [title, status, trace_id]
      properties:
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        trace_id: { type: string }
    ConflictProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties: { current_sequence: { type: integer } }
    MatchSide:
      type: object
      required: [display_name]
      properties:
        team_id: { type: [string, 'null'], format: uuid }
        display_name: { type: string, minLength: 1, maxLength: 100 }
...
    Match:
      type: object
      required: [id, organization_id, status, event_sequence, home, away]
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        status: { type: string, enum: [DRAFT, READY, LIVE, PAUSED, COMPLETED, FINALIZED, ABANDONED] }
        event_sequence: { type: integer, minimum: 0 }
        home: { $ref: '#/components/schemas/MatchSide' }
        away: { $ref: '#/components/schemas/MatchSide' }
...
    EventSubject:
      type: object
      required: [role, participant_id]
      properties:
        role: { type: string, enum: [PRIMARY, SCORER, ASSISTER, OPPONENT, PLAYER_ON, PLAYER_OFF] }
        participant_id: { type: string, format: uuid }
    AppendEventRequest:
      type: object
      required: [client_event_id, expected_sequence, action_code, side, subjects]
      properties:
        client_event_id: { type: string, format: uuid }
        expected_sequence: { type: integer, minimum: 0 }
        action_code: { type: string, enum: [GOAL, ASSIST, SHOT_ON_TARGET, TACKLE_WON, INTERCEPTION, GOALKEEPER_SAVE, FOUL_COMMITTED, FOUL_WON, SUBSTITUTION, SCORE_ADJUSTMENT] }
        side: { type: string, enum: [HOME, AWAY] }
        subjects: { type: array, minItems: 1, items: { $ref: '#/components/schemas/EventSubject' } }
        qualifiers: { type: object, additionalProperties: true }
...
    MatchEvent:
      type: object
      required: [id, sequence, action_code, side, subjects]
      properties:
        id: { type: string, format: uuid }
        sequence: { type: integer }
        action_code: { type: string }
        side: { type: string, enum: [HOME, AWAY] }
        subjects: { type: array, items: { $ref: '#/components/schemas/EventSubject' } }
    EventList:
      type: object
      required: [items, last_sequence]
      properties:
        last_sequence: { type: integer }
        items: { type: array, items: { $ref: '#/components/schemas/MatchEvent' } }
    Snapshot:
      type: object
      required: [match_id, event_sequence, home_score, away_score, status]
      properties:
        match_id: { type: string, format: uuid }
        event_sequence: { type: integer }
        home_score: { type: integer }
        away_score: { type: integer }
        status: { type: string }
    AppendEventResponse:
      type: object
      required: [event, snapshot]
      properties:
        event: { $ref: '#/components/schemas/MatchEvent' }
        snapshot: { $ref: '#/components/schemas/Snapshot' }